UCS/SYS: rename sys/socket.h to sys/sock.h
