+++ src/ucp/core/ucp_ep.c
@@ -19,7 +19,7 @@
 #include <ucs/debug/memtrack.h>
 #include <ucs/debug/log.h>
 #include <ucs/sys/string.h>
-#include <ucs/sys/socket.h>
+#include <ucs/sys/sock.h>
 #include <string.h>
