+++ src/ucp/wireup/select.c
@@ -9,7 +9,7 @@
 
 #include <ucs/algorithm/qsort_r.h>
 #include <ucs/datastruct/queue.h>
-#include <ucs/sys/socket.h>
+#include <ucs/sys/sock.h>
 #include <ucp/core/ucp_ep.inl>
 #include <string.h>
 #include <inttypes.h>
