+++ src/ucp/wireup/wireup_ep.c
@@ -12,7 +12,7 @@
 #include <ucs/arch/atomic.h>
 #include <ucs/datastruct/queue.h>
 #include <ucs/type/class.h>
-#include <ucs/sys/socket.h>
+#include <ucs/sys/sock.h>
 #include <ucp/core/ucp_ep.inl>
 #include <ucp/core/ucp_request.inl>
