+++ src/uct/ib/rdmacm/rdmacm_md.h
@@ -8,7 +8,7 @@
 
 #include "rdmacm_def.h"
 #include <uct/base/uct_md.h>
-#include <ucs/sys/socket.h>
+#include <ucs/sys/sock.h>
 #include <ucs/time/time.h>
 #include <rdma/rdma_cma.h>
