+++ test/gtest/Makefile.am
@@ -149,7 +149,7 @@ gtest_SOURCES = \
 	ucs/test_strided_alloc.cc \
 	ucs/test_string.cc \
 	ucs/test_sys.cc \
-	ucs/test_socket.cc \
+	ucs/test_sock.cc \
 	ucs/test_time.cc \
 	ucs/test_twheel.cc \
 	ucs/test_frag_list.cc \
