+++ test/gtest/ucs/test_sys.cc
@@ -8,7 +8,7 @@
 extern "C" {
 #include <ucs/sys/module.h>
 #include <ucs/sys/sys.h>
-#include <ucs/sys/socket.h>
+#include <ucs/sys/sock.h>
 #include <ucs/type/spinlock.h>
 #include <ucs/time/time.h>
 }
